+++ unit_tests/engine/geometry_string.cpp
@@ -1,4 +1,4 @@
-#include <boost/test/floating_point_comparison.hpp>
+#include <boost/test/tools/floating_point_comparison.hpp>
 #include <boost/test/unit_test.hpp>
 
 #include "engine/polyline_compressor.hpp"